requirements.txt
attrs==25.4.0
genson==1.3.0
json5==0.12.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.28.0
